riski-extractor/pyproject.toml
[project]
name = "riski-extractor"
version = "0.0.1a1"
description = "Python application for extracting information from munich's political information system RIS"
readme = "README.md"
requires-python = "==3.12.11"
license = { file = "LICENSE" }
authors = [
    { name = "Fabian Reinold", email = "fabian.reinold@muenchen.de" },
    { name = "Stefan Hentschel", email = "stefan.hentschel@muenchen.de" },
    { name = "Annalena Wittmann", email = "annalena.wittmann@muenchen.de" },
]
dependencies = [
    "gitpython==3.1.45",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-docling==1.0.0",
    "langchain-openai==0.3.29",
    "pydantic[email]==2.11.7",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",
    "stamina==25.1.0",
    "tokenizers==0.21.2",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "truststore==0.10.3",
]

[project.optional-dependencies]
pgvector = [
    "asyncpg==0.30.0",
    "langchain-postgres==0.0.15",
    "psycopg[binary,pool]==3.2.9",
]

[dependency-groups]
dev = ["ipykernel==6.30.1", "pre-commit==4.2.0", "ruff==0.12.5", "ty==0.0.1a17"]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true